auto_funny_videos/requirements.txt
requests
yt-dlp
ffmpeg-python
openai
google-api-python-client
google-auth
python-dotenv  # If you want to load environment variables from a .env file
